requirements.txt
kafka==1.3.5
python-dotenv==0.10.3
